+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * 【core   】      修复Convert.digitToChinese(0)输出金额无`元整问题`（issue#3662@Github）
 * 【core   】      修复CsvParser中对正文中双引号处理逻辑问题（pr#1244@Gitee）
 * 【core   】      修复ZipUtil.zip压缩到本目录时可能造成的死循环问题（issue#IAGYDG@Gitee）
+* 【cache  】      修复AbstractCache.get中锁不一致导致的并发问题（issue#3686@Github）
 
 -------------------------------------------------------------------------------------------------------------
 # 5.8.29(2024-07-03)